% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Function for print a message}
\usage{
hello(message1, message2)
}
\arguments{
\item{a}{a message}

\item{b}{a message}
}
\value{
Print word \code{a} and \code{b}
}
\description{
Function for print a message
}
\examples{
hello('Hello', 'world')

}
