% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momos.R
\name{momos}
\alias{momos}
\title{The MOMOS function}
\usage{
momos(params = NULL)
}
\arguments{
\item{params}{params of the function}
}
\value{
Table with the data
}
\description{
The MOMOS function
}
\examples{
params <- list()
params$Necromasa <- 2140
params$HLo <- 2250
params$HSo <- 19150
params$Co <- 55.40
params$Kvl <- 0.2070
params$Kvs <- 0.00057
params$fs <- 0.00002
params$Khl <- 0.0638
params$Khs <- 0.00077
params$Khls <- 0.1581
params$Kmb <- 0.001
params$Kresp <- 0.1419
params$Ci <- 50.04
params$from <- 0
params$to <- 30
params$by <- 1
out <- momos(params)

}
