% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world.R
\name{world}
\alias{world}
\title{Function for print other message}
\usage{
world(message1, message2)
}
\arguments{
\item{a}{a message}

\item{b}{a message}
}
\value{
Print word \code{a} and \code{b}
}
\description{
Function for print other message
}
\examples{
world('Hello', 'world')

}
