% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momos.R
\name{momos}
\alias{momos}
\title{The MOMOS function}
\usage{
momos(pars)
}
\arguments{
\item{params}{params of the function}
}
\value{
Table with the data
}
\description{
The MOMOS function
}
\examples{
out <- momos(pars)

}
